web: unify overview + sidebar trigger button logic
