+++ web/src/OverviewTableTriggerButton.tsx
@@ -1,13 +1,8 @@
-import React from "react"
 import styled from "styled-components"
-import { Tags } from "./analytics"
-import { ReactComponent as TriggerButtonSvg } from "./assets/svg/trigger-button.svg"
-import { InstrumentedButton } from "./instrumentedComponents"
 import { AnimDuration, Color, mixinResetButtonStyle } from "./style-helpers"
-import { triggerTooltip, triggerUpdate } from "./trigger"
-import { TriggerMode } from "./types"
+import TriggerButton from "./TriggerButton"
 
-export let TriggerButtonRoot = styled(InstrumentedButton)`
+export const OverviewTableTriggerButton = styled(TriggerButton)`
   ${mixinResetButtonStyle};
   display: flex;
   align-items: center;
@@ -46,63 +41,3 @@ export let TriggerButtonRoot = styled(InstrumentedButton)`
     fill: ${Color.blue};
   }
 `
-
-type TriggerButtonProps = {
-  isBuilding: boolean
-  hasBuilt: boolean
-  triggerMode: TriggerMode
-  hasPendingChanges: boolean
-  isQueued: boolean
-  resourceName: string
-  analyticsTags: Tags
-  onTrigger: () => void
-}
-
-function OverviewTableTriggerButton(props: TriggerButtonProps) {
-  let isManual =
-    props.triggerMode === TriggerMode.TriggerModeManual ||
-    props.triggerMode === TriggerMode.TriggerModeManualWithAutoInit
-  let isAutoInit =
-    props.triggerMode === TriggerMode.TriggerModeAuto ||
-    props.triggerMode === TriggerMode.TriggerModeManualWithAutoInit
-
-  // clickable (i.e. trigger button will appear) if it doesn't already have some kind of pending / active build
-  let clickable =
-    !props.isQueued && // already queued for manual run
-    !props.isBuilding && // currently building
-    !(isAutoInit && !props.hasBuilt) // waiting to perform its initial build
-
-  let isEmphasized = false
-  if (clickable) {
-    if (props.hasPendingChanges && isManual) {
-      isEmphasized = true
-    } else if (!props.hasBuilt && !isAutoInit) {
-      isEmphasized = true
-    }
-  }
-
-  let classes = []
-  if (!clickable) {
-    classes.push("is-disabled")
-  }
-  if (props.isBuilding) {
-    classes.push("is-building")
-  }
-  if (isEmphasized) {
-    classes.push("is-emphasized")
-  }
-  return (
-    <TriggerButtonRoot
-      aria-disabled={!clickable}
-      onClick={() => triggerUpdate(props.resourceName)}
-      className={classes.join(" ")}
-      title={triggerTooltip(clickable, isEmphasized, props.isQueued)}
-      analyticsName={"ui.web.triggerResource"}
-      analyticsTags={props.analyticsTags}
-    >
-      <TriggerButtonSvg />
-    </TriggerButtonRoot>
-  )
-}
-
-export default React.memo(OverviewTableTriggerButton)
